requirements.txt
absl-py==0.13.0
aio-pika==6.8.0
aiofiles==0.7.0
aiohttp==3.7.4
aiormq==3.3.1
alabaster==0.7.12
alembic==1.4.3
appdirs==1.4.4
APScheduler==3.7.0
argostranslate==1.1.2
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bidict==0.21.2
boto3==1.18.36
botocore==1.21.36
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==1.6.0
colorclass==2.2.0
coloredlogs==15.0.1
colorhash==1.0.3
croniter==1.0.15
cryptography==3.4.8
ctranslate2==1.17.1
cycler==0.10.0
decorator==4.4.2
Django==4.1
django-lint==2.0.4
djangorestframework==3.13.1
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
expiringdict==1.2.1
fbmessenger==6.0.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-HTTPAuth==4.5.0
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
Flask-WTF==0.15.1
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.39.0
gunicorn==20.1.0
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.19.1
httptools==0.3.0
httpx==0.15.4
humanfriendly==9.2
idna==3.3
imagesize==1.2.0
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langdetect==1.0.8
libretranslate==1.1.0
limits==1.5.1
linecache2==1.0.0
logilab-astng==0.24.3
logilab-common==1.9.7
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mattermostwrapper==2.2
mock==4.0.3
Morfessor==2.0.6
multidict==4.7.6
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pamqp==2.3.0
pika==1.2.0
Pillow==8.3.2
pluggy==1.0.0
polyglot==16.7.4
prompt-toolkit==2.0.10
protobuf==3.19.0
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld2==0.41
pycparser==2.20
pydot==1.4.2
Pygments==2.10.0
PyJWT==2.0.0
pykwalify==1.8.0
pylint==0.28.0
pymongo==3.10.1
PyMySQL==1.0.2
pyparsing==3.0.1
PyQt5==5.15.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
pyTelegramBotAPI==3.8.3
pytest==6.2.5
pytextbelt==0.5
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2022.1
PyYAML==5.4.1
questionary==1.5.2
rasa==2.8.4
rasa-sdk==2.8.1
rasa-x==0.42.1
redis==3.5.3
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.16.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.6.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.2
scipy==1.7.1
sentencepiece==0.1.91
sentry-sdk==1.2.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.26
sqlparse==0.4.2
stanza==1.1.1
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.4
tensorflow-addons==0.13.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.12.0
tensorflow-probability==0.13.0
tensorflow-text==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.2.0
toml==0.10.2
torch==1.9.1
tqdm==4.62.2
traceback2==1.4.0
translatehtml==1.5.1
twilio==6.50.1
typeguard==2.12.1
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==2.1
ujson==1.35
unittest-xml-reporting==3.0.2
unittest2==1.1.0
urllib3==1.26.7
uvloop==0.14.0
waitress==1.4.4
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==2.0.2
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
